libstormir: Untested work on prealloc layer and others
Next step is actually testing the backend and fixing issues with it.

+++ libstormir/src/lir/mod.rs
@@ -1,6 +1,6 @@
 pub mod shim;
 
-pub type LIRVarId = usize;
+pub type LIRVarId = u64;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum LIRInstruction {
